CSS: Make show/hide/toggle methods a module
Fixes #2193

Unit test changes some uses of .show() and .hide() to .css( "display", ... ),
there was already an implicit assumption in several of the existing tests.
