+++ src/css/showHide.js
@@ -1,6 +1,7 @@
 define( [
-	"../data/var/dataPriv"
-], function( dataPriv ) {
+	"../data/var/dataPriv",
+	"../css/var/isHidden"
+], function( dataPriv, isHidden ) {
 
 function showHide( elements, show ) {
 	var display, elem,
@@ -43,6 +44,26 @@ function showHide( elements, show ) {
 	return elements;
 }
 
-return showHide;
+jQuery.fn.extend( {
+	show: function() {
+		return showHide( this, true );
+	},
+	hide: function() {
+		return showHide( this );
+	},
+	toggle: function( state ) {
+		if ( typeof state === "boolean" ) {
+			return state ? this.show() : this.hide();
+		}
+
+		return this.each( function() {
+			if ( isHidden( this ) ) {
+				jQuery( this ).show();
+			} else {
+				jQuery( this ).hide();
+			}
+		} );
+	}
+} );
 
 } );
