+++ test/unit/queue.js
@@ -231,6 +231,8 @@ QUnit.asyncTest( "fn.promise( \"queue\" ) - called whenever last queue function 
 	foo.dequeue( "queue" );
 } );
 
+if ( jQuery.fn.animate ) {
+	
 QUnit.asyncTest( "fn.promise( \"queue\" ) - waits for animation to complete before resolving", 2, function( assert ) {
 	var foo = jQuery( "#foo" ),
 		test = 1;
@@ -251,6 +253,7 @@ QUnit.asyncTest( "fn.promise( \"queue\" ) - waits for animation to complete befo
 	} );
 
 } );
+}
 
 QUnit.test( ".promise(obj)", function( assert ) {
 	assert.expect( 2 );
